// amplify/backend/function/jobUpdateOwner/src/API.ts
export type UpdateOwnerResponseType = {
  __typename: "UpdateOwnerResponseType",
  isSuccess: boolean,
  message: string,
  result?: string | null,
};
// ...
export enum EmployeeType {
  OTHER = "OTHER",
  NORMAL = "NORMAL",
  MANAGER = "MANAGER",
  SUPER_MANAGER = "SUPER_MANAGER",
}


export enum BooleanType {
  TRUE = "TRUE",
  FALSE = "FALSE",
}


export type Employee = {
  __typename: "Employee",
  username: string,
  companyID: string,
  no: string,
  groupID: string,
  superiorUsername: string,
  firstName: string,
  lastName: string,
  grade: string,
  email: string,
  sub?: string | null,
  manager: EmployeeType,
  isCompanyAdmin?: boolean | null,
  isDeleted: BooleanType,
  createdAt: string,
  updatedAt: string,
  company?: Company | null,
  superior?: Employee | null,
  group?: Group | null,
};

export type Company = {
  __typename: "Company",
  id: string,
  name: string,
  startMonth: number,
  shortName?: string | null,
  url?: string | null,
  isContractEvaluation?: boolean | null,
  isContractReport?: boolean | null,
  createdAt: string,
  updatedAt: string,
};

export type Group = {
  __typename: "Group",
  id: string,
  companyID: string,
  no: string,
  name: string,
  createdAt: string,
  updatedAt: string,
};
// ...
export type Sheet = {
  __typename: "Sheet",
  id: string,
  sub: string,
  year: number,
  companyID: string,
  groupID: string,
  grade: string,
  careerPlan?: string | null,
  careerPlanComment?: string | null,
  reviewComment?: string | null,
  reviewDate?: string | null,
  selfCheckDate?: string | null,
  firstComment?: string | null,
  firstCheckDate?: string | null,
  secondComment?: string | null,
  secondCheckDate?: string | null,
  overAllEvaluation?: number | null,
  statusValue?: number | null,
  interviewPlanDate?: string | null,
  interviewPlanComment?: string | null,
  InterviewMid1Date?: string | null,
  InterviewMid1Comment?: string | null,
  InterviewMid2Date?: string | null,
  InterviewMid2Comment?: string | null,
  InterviewMid3Date?: string | null,
  InterviewMid3Comment?: string | null,
  revieweeUsername?: string | null,
  secondUsername?: string | null,
  topUsername?: string | null,
  sheetGroupName?: string | null,
  referencer?: Array< string > | null,
  reviewee: string,
  topReviewers?: Array< string > | null,
  secondReviewers?: Array< string > | null,
  createdAt: string,
  updatedAt: string,
  revieweeEmployee?: Employee | null,
  secondEmployee?: Employee | null,
  topEmployee?: Employee | null,
  group?: Group | null,
  section?: ModelSectionConnection | null,
};

export type ModelSectionConnection = {
  __typename: "ModelSectionConnection",
  items:  Array<Section | null >,
  nextToken?: string | null,
};

export type Section = {
  __typename: "Section",
  id: string,
  sheetID: string,
  no: string,
  sectionCategoryName?: string | null,
  reviewee?: string | null,
  topReviewers?: Array< string > | null,
  secondReviewers?: Array< string > | null,
  referencer?: Array< string > | null,
  createdAt: string,
  updatedAt: string,
  objective?: ModelObjectiveConnection | null,
};

export type ModelObjectiveConnection = {
  __typename: "ModelObjectiveConnection",
  items:  Array<Objective | null >,
  nextToken?: string | null,
};

export type Objective = {
  __typename: "Objective",
  id: string,
  sectionID: string,
  createdAt: string,
  companyID: string,
  content: string,
  result?: string | null,
  priority?: string | null,
  selfEvaluation?: number | null,
  firstEvaluation?: number | null,
  lastEvaluation?: number | null,
  progress?: number | null,
  expStartDate?: string | null,
  expDoneDate?: string | null,
  reviewee?: string | null,
  topReviewers?: Array< string > | null,
  secondReviewers?: Array< string > | null,
  referencer?: Array< string > | null,
  updatedAt: string,
};
// ...
export enum ReportWorkingStatus {
  OK = "OK",
  InTask = "InTask",
  InProblem = "InProblem",
}


export type Report = {
  __typename: "Report",
  id: string,
  sub: string,
  date: string,
  groupID?: string | null,
  reviewee?: string | null,
  reviewer?: Array< string | null > | null,
  companyID?: string | null,
  referencer?: Array< string | null > | null,
  revieweeComments?: RevieweeReportComment | null,
  reviewerComments?: ReviewerReportComment | null,
  workStatus?: ReportWorkingStatus | null,
  createdAt: string,
  updatedAt: string,
  revieweeEmployee?: Employee | null,
};

export type RevieweeReportComment = {
  __typename: "RevieweeReportComment",
  work?: string | null,
  other?: string | null,
  status?: string | null,
};

export type ReviewerReportComment = {
  __typename: "ReviewerReportComment",
  superior?: string | null,
};